refactor : MemberService 수정 (#18)
- member 정보를 수정할 때,
받은 이미지가 null인지를 파악한 후에 null 이 아니라면 S3에 파일을 업로드 하도록 수정
- Member 정보를 조회할 때,
Pet의 정보도 함게 조회하도록 수정

+++ src/main/java/com/me/wodada/member/service/MemberService.java
@@ -5,12 +5,14 @@ import com.me.wodada.member.domain.Member;
 import com.me.wodada.member.dto.request.MemberInfoUpdateReq;
 import com.me.wodada.member.dto.response.MemberInfoRes;
 import com.me.wodada.member.repository.MemberRepository;
+import com.me.wodada.pet.dto.response.PetInfoRes;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional(readOnly = true)
@@ -21,26 +23,28 @@ public class MemberService {
     private final AmazonS3Service amazonS3Service;
 
     @Transactional
-    public void updateMemberInfo(Member member, MemberInfoUpdateReq memberInfoUpdateReq, List<MultipartFile> images) {
-        if (member.getProfileImageUrl() != null){
-            amazonS3Service.deleteFile(member.getProfileImageUrl());
-        }
-
-        String profileUrl = amazonS3Service.uploadFile(images).get(0);
-        updateMember(member, memberInfoUpdateReq, profileUrl);
-
-    }
-
-    private void updateMember(Member member, MemberInfoUpdateReq memberInfoUpdateReq, String profileUrl) {
+    public void updateMemberInfo(Member member, MemberInfoUpdateReq request, MultipartFile image) {
         // TODO : Exception 처리
         Member findMember = memberRepository.findByEmailAndProvider(member.getEmail(), member.getProvider())
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 유저입니다."));
 
-        findMember.updateProfile(memberInfoUpdateReq, profileUrl);
-        // TODO : Locaiont update -> Role User로 변경
+        if (image != null){
+            if (member.getProfileImageUrl() != null){
+                amazonS3Service.deleteFile(member.getProfileImageUrl());
+            }
+            String profileUrl = amazonS3Service.uploadFile(List.of(image)).get(0);
+            findMember.addProfileImage(profileUrl);
+        }
+        findMember.updateProfile(request);
     }
 
-    public MemberInfoRes getMemberInfo(Member member) {
-        return MemberInfoRes.from(member);
+    public MemberInfoRes getMemberInfo(Long memberId) {
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 멤버입니다"));
+
+        List<PetInfoRes> petInfoResList = member.getPetList().stream()
+                .map(pet -> PetInfoRes.from(pet))
+                .collect(Collectors.toList());
+        return MemberInfoRes.from(member, petInfoResList);
     }
 }
